// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/models/tapp-qualification-support-staff.model.ts
export class TappSupportStuffModel {

    id: string;
    uuid: string;
    projectConceptUuid: string;
    projectConceptId: number;
    gobFund: number;
    rpaFund: number;
    dpaFund: number;
    others: number;
    
    

   // dppModeOfFinancing: any;
   // currencyRateList: DppCurrencyRateModel[] = new Array<DppCurrencyRateModel>();
}